Move replica planner kernels to Triton transport module
