+++ megatron/core/transformer/moe/replica_planner.py
@@ -51,8 +51,15 @@ from megatron.core.fp8_utils import is_mxfp8tensor
 from megatron.core.transformer.moe.replica_weight_triton import (
     MAX_REPLICA_WEIGHT_SMS,
     compile_replica_weight_kernels,
+    launch_compact_routing_map,
     launch_replica_grad_reduce,
+    launch_replica_placement,
+    launch_replica_route_mapping,
+    launch_replica_route_ranking,
     launch_replica_weight_prefetch,
+)
+from megatron.core.transformer.moe.replica_weight_triton import (
+    planner_route_partition_count as _route_partition_count,
 )
 from megatron.core.utils import nvtx_decorator
 
@@ -61,41 +68,10 @@ try:
 except ImportError:
     get_dummy_wgrad = None
 
-try:
-    import triton
-    import triton.language as tl
-
-    HAVE_TRITON = True
-except ImportError:
-    from unittest.mock import MagicMock
-
-    from megatron.core.utils import null_decorator
-
-    triton = MagicMock()
-    triton.jit = null_decorator
-    tl = MagicMock()
-    HAVE_TRITON = False
-
-
-# Grid width for the planner's persistent and cooperative launches. It stays
-# below the resident-block limit of every supported device while keeping enough
-# programs to saturate one.
-_MAX_PLANNER_PROGRAMS = 128
-
 
 def _next_power_of_two(value: int) -> int:
     """Round a positive dimension up to the power of two Triton tiles use."""
     return 1 << (value - 1).bit_length()
-
-
-def _route_partition_count(num_routes: int) -> int:
-    """Return the route partitioning shared by route ranking and route mapping.
-
-    Route ranking uses a cooperative launch, which must keep the complete grid
-    resident. The default is safe on the validated GB300 workload and retains
-    substantially more parallelism than a conservative 64-program setting.
-    """
-    return min(_MAX_PLANNER_PROGRAMS, num_routes)
 
 
 def _discard_runtime_parameter_grad(parameter: torch.nn.Parameter) -> None:
@@ -142,9 +118,7 @@ class ReplicaPlannerWorkspace:
     # local routes to expert on source; allocation[expert, destination] is the
     # final partition of that expert's global route stream across ranks.
     gathered_counts: torch.Tensor
-    expert_totals: torch.Tensor
     balance: torch.Tensor
-    receiver_quotas: torch.Tensor
     allocation: torch.Tensor
     placement_grid_sync: torch.Tensor
     # Per-expert destination segment ends, rebased into this rank's local
@@ -209,18 +183,16 @@ class ReplicaPlannerWorkspace:
             ep_size=ep_size,
             num_local_experts=num_local_experts,
             gathered_counts=torch.empty((ep_size, num_experts), **int32),
-            expert_totals=torch.empty(num_experts, **int32),
             balance=torch.empty(ep_size, **int32),
-            receiver_quotas=torch.empty((ep_size, ep_size), **int32),
             allocation=torch.empty((num_experts, ep_size), **int32),
-            placement_grid_sync=torch.zeros(2, **int32),
+            placement_grid_sync=torch.zeros(1, **int32),
             destination_boundaries=torch.empty((num_experts, _next_power_of_two(ep_size)), **int32),
             expert_replica_slots=torch.empty((num_experts, ep_size), **int32),
             sort_route_metadata=torch.empty(num_routes, **int32),
             sort_partition_counts=torch.empty(
                 (_route_partition_count(num_routes), num_experts), **int32
             ),
-            sort_grid_sync=torch.zeros(2, **int32),
+            sort_grid_sync=torch.zeros(1, **int32),
             sort_stream=torch.cuda.Stream(device=device),
             virtual_experts=torch.empty(
                 (num_tokens, router_topk), dtype=torch.int64, device=device
@@ -1624,292 +1596,6 @@ def wait_replica_grad_reduce_after_dispatch_backward(
     return _ReplicaWaitGradReduce.apply(hidden_states, *bridge.source_parameters, bridge, plan)
 
 
-@triton.jit
-def _cooperative_grid_barrier(grid_sync, num_programs):
-    """Synchronize a cooperatively launched Triton grid."""
-    generation = tl.atomic_add(grid_sync + 1, 0, sem="acquire", scope="gpu")
-    arrival = tl.atomic_add(grid_sync, 1, sem="acq_rel", scope="gpu")
-    if arrival == num_programs - 1:
-        tl.atomic_xchg(grid_sync, 0, sem="release", scope="gpu")
-        tl.atomic_add(grid_sync + 1, 1, sem="acq_rel", scope="gpu")
-    else:
-        observed_generation = generation
-        while observed_generation == generation:
-            observed_generation = tl.load(grid_sync + 1, cache_modifier=".cv", volatile=True)
-        tl.atomic_add(grid_sync + 1, 0, sem="acquire", scope="gpu")
-
-
-@triton.jit
-def _plan_replica_placement_kernel(
-    gathered_tokens_per_expert,
-    global_tokens_per_expert,
-    rank_load_balance,
-    receiver_quotas,
-    expert_rank_allocations,
-    destination_boundaries,
-    experts_to_copy,
-    expert_replica_slots,
-    grid_sync,
-    RANK_ROUTE_CAPACITY: tl.constexpr,
-    SOURCE_EP_RANK: tl.constexpr,
-    EP_SIZE: tl.constexpr,
-    NUM_EXPERTS: tl.constexpr,
-    NUM_EXPERTS_PER_GPU: tl.constexpr,
-    BLOCK_EP_SIZE: tl.constexpr,
-    BLOCK_NUM_EXPERTS_PER_GPU: tl.constexpr,
-    BLOCK_NUM_EXPERTS: tl.constexpr,
-):
-    """Compute the whole deterministic placement in one cooperative launch.
-
-    Program ``rank`` owns the semantic experts of one EP rank and the replica
-    slots of that same rank. It runs the four placement steps in order, with a
-    grid barrier wherever a step needs a value another program produced:
-
-    1. Sum the gathered histograms for its native experts and publish the
-       resulting group balance relative to one rank's route capacity.
-    2. Replay the deterministic rank-pairing greedy over all balances. Every
-       program derives the identical quota matrix, so only the row this program
-       owns has to be kept and stored.
-    3. Split its own quotas across its native experts. The allocation rows this
-       step rewrites belong to this program alone, so the whole
-       ``[expert, destination]`` tile stays in registers.
-    4. After the allocation is globally visible, rank the remote experts routed
-       to this program and record them in its replica slots.
-
-    Step 3 also emits the route mapper's search table. Each expert's routes
-    form one global stream ordered by source rank, so subtracting the routes
-    contributed by lower source ranks rebases the destination segment ends
-    into the local ordinal space the mapper works in.
-
-    Keeping the steps in one kernel removes four launches and, more
-    importantly, keeps the quota vector and the allocation tile in registers
-    instead of round-tripping them through global memory once per greedy step.
-    The greedy loops stay rolled: their bodies work on whole tiles, so
-    unrolling them multiplies the instruction footprint without exposing any
-    parallelism the sequential dependency would allow.
-    """
-    rank = tl.program_id(0)
-    ranks = tl.arange(0, BLOCK_EP_SIZE)
-    valid_ranks = ranks < EP_SIZE
-    local_experts = tl.arange(0, BLOCK_NUM_EXPERTS_PER_GPU)
-    valid_local_experts = local_experts < NUM_EXPERTS_PER_GPU
-    native_experts = rank * NUM_EXPERTS_PER_GPU + local_experts
-
-    # Step 1: global route totals for this rank's native experts, and the
-    # amount by which their sum exceeds one rank's route capacity.
-    source_counts = tl.load(
-        gathered_tokens_per_expert + ranks[:, None] * NUM_EXPERTS + native_experts[None, :],
-        mask=valid_ranks[:, None] & valid_local_experts[None, :],
-        other=0,
-    )
-    native_totals = tl.sum(source_counts, axis=0).to(tl.int32)
-    routes_before_source = tl.sum(
-        tl.where(ranks[:, None] < SOURCE_EP_RANK, source_counts, 0), axis=0
-    ).to(tl.int32)
-    tl.store(global_tokens_per_expert + native_experts, native_totals, mask=valid_local_experts)
-    tl.store(
-        rank_load_balance + rank,
-        tl.sum(native_totals, axis=0).to(tl.int32) - RANK_ROUTE_CAPACITY,
-    )
-
-    _cooperative_grid_barrier(grid_sync, EP_SIZE)
-
-    # Step 2: greedily pair the lowest-index most-loaded and roomiest ranks.
-    # Equal balances choose the lowest rank, making the quota matrix identical
-    # on every EP rank. quotas[owner, receiver] records how many routes the
-    # receiver should take from that owner's experts; it does not choose the
-    # experts yet.
-    balances = tl.load(rank_load_balance + ranks, mask=valid_ranks, other=0)
-    quotas = tl.zeros((BLOCK_EP_SIZE,), dtype=tl.int32)
-    for _ in tl.range(0, EP_SIZE, 1, loop_unroll_factor=1):
-        maximum = tl.max(tl.where(valid_ranks, balances, -2147483648), axis=0)
-        minimum = tl.min(tl.where(valid_ranks, balances, 2147483647), axis=0)
-        overloaded = tl.min(
-            tl.where(valid_ranks & (balances == maximum), ranks, BLOCK_EP_SIZE), axis=0
-        )
-        receiver = tl.min(
-            tl.where(valid_ranks & (balances == minimum), ranks, BLOCK_EP_SIZE), axis=0
-        )
-        active = maximum > 0
-        moved = tl.where(active, -minimum, 0).to(tl.int32)
-        quotas = tl.where(active & (overloaded == rank) & (ranks == receiver), moved, quotas)
-        balances = tl.where(active & (ranks == overloaded), balances - moved, balances)
-        balances = tl.where(active & (ranks == receiver), 0, balances)
-    tl.store(receiver_quotas + rank * EP_SIZE + ranks, quotas, mask=valid_ranks)
-
-    # Step 3: assign this rank's receive quotas to its largest expert segments.
-    # Repeatedly pair the largest outstanding receiver quota with the largest
-    # remaining expert segment. Expert ties choose the lowest local expert id,
-    # matching the reference placement policy. A semantic expert may therefore
-    # be split between its owner and one or more replica destinations.
-    # allocations[local_expert, destination] starts with the native owner
-    # executing every route and is migrated away from there in place.
-    remaining = native_totals
-    allocations = tl.where(ranks[None, :] == rank, native_totals[:, None], 0)
-    for _ in tl.range(0, EP_SIZE + NUM_EXPERTS_PER_GPU, 1, loop_unroll_factor=1):
-        max_quota = tl.max(tl.where(valid_ranks, quotas, -1), axis=0)
-        destination = tl.min(
-            tl.where(valid_ranks & (quotas == max_quota), ranks, BLOCK_EP_SIZE), axis=0
-        )
-        max_remaining = tl.max(tl.where(valid_local_experts, remaining, -1), axis=0)
-        local_expert = tl.min(
-            tl.where(
-                valid_local_experts & (remaining == max_remaining),
-                local_experts,
-                BLOCK_NUM_EXPERTS_PER_GPU,
-            ),
-            axis=0,
-        )
-        active = max_quota > 0
-        moved = tl.where(active, tl.minimum(max_quota, max_remaining), 0).to(tl.int32)
-        # One pass over the tile: the migrated routes leave the owner column
-        # and arrive in the destination column of the same expert row.
-        transfer = tl.where(
-            ranks[None, :] == destination, moved, tl.where(ranks[None, :] == rank, -moved, 0)
-        )
-        allocations += tl.where((local_experts[:, None] == local_expert) & active, transfer, 0)
-        remaining = tl.where(active & (local_experts == local_expert), remaining - moved, remaining)
-        quotas = tl.where(active & (ranks == destination), quotas - moved, quotas)
-    tl.store(
-        expert_rank_allocations + native_experts[:, None] * EP_SIZE + ranks[None, :],
-        allocations,
-        mask=valid_local_experts[:, None] & valid_ranks[None, :],
-    )
-    tl.store(
-        destination_boundaries + native_experts[:, None] * BLOCK_EP_SIZE + ranks[None, :],
-        tl.cumsum(allocations, axis=1) - routes_before_source[:, None],
-        mask=valid_local_experts[:, None],
-    )
-
-    _cooperative_grid_barrier(grid_sync, EP_SIZE)
-
-    # Step 4: fill this rank's replica slots by count descending, expert id
-    # descending. Only remote experts with a positive allocation need copied
-    # weights. The ordering is observable because it defines replica-slot ids,
-    # so the expert id tie-break deliberately runs in descending order for
-    # deterministic parity. Unused slots are filled with -1.
-    experts = tl.arange(0, BLOCK_NUM_EXPERTS)
-    owner = experts // NUM_EXPERTS_PER_GPU
-    valid_remote = (experts < NUM_EXPERTS) & (owner != rank)
-    counts = tl.load(
-        expert_rank_allocations + experts * EP_SIZE + rank, mask=valid_remote, other=-1
-    )
-    for slot in tl.range(0, NUM_EXPERTS_PER_GPU, 1, loop_unroll_factor=1):
-        maximum = tl.max(tl.where(valid_remote, counts, -1), axis=0)
-        expert = tl.max(tl.where(valid_remote & (counts == maximum), experts, -1), axis=0)
-        selected = tl.where(maximum > 0, expert, -1).to(tl.int32)
-        tl.store(experts_to_copy + rank * NUM_EXPERTS_PER_GPU + slot, selected)
-        tl.store(expert_replica_slots + selected * EP_SIZE + rank, slot, mask=selected >= 0)
-        counts = tl.where(experts == expert, -1, counts)
-
-
-@triton.jit
-def _rank_routes_within_experts_kernel(
-    flat_topk_indices,
-    route_metadata,
-    partition_counts,
-    grid_sync,
-    NUM_ROUTES: tl.constexpr,
-    NUM_EXPERTS: tl.constexpr,
-    BLOCK_NUM_EXPERTS: tl.constexpr,
-    BLOCK_NUM_ROUTES: tl.constexpr,
-    BLOCK_SCAN_PARTITIONS: tl.constexpr,
-    NUM_SCAN_EXPERTS: tl.constexpr,
-):
-    """Give every route its stable ordinal within its expert's local stream.
-
-    A cooperative grid splits the original route stream into consecutive
-    partitions. The first pass computes each route's partition-local ordinal
-    and packs it with the expert id in one int32 scratch value. Programs then
-    publish their histograms, synchronize, and prefix-scan the partition
-    counts. A route's local ordinal is therefore
-    ``partition_counts[partition, expert] + partition_local_ordinal``, which is
-    exactly its position within the expert bucket a stable sort would produce.
-    The route mapper reconstructs it from the same partitioning, so no sorted
-    index array is ever materialized.
-    """
-    partition = tl.program_id(0)
-    num_partitions = tl.num_programs(0)
-    expert_offsets = tl.arange(0, BLOCK_NUM_EXPERTS)
-    valid_experts = expert_offsets < NUM_EXPERTS
-    routes_per_partition = tl.cdiv(NUM_ROUTES, num_partitions)
-    partition_start = partition * routes_per_partition
-    partition_end = tl.minimum(partition_start + routes_per_partition, NUM_ROUTES)
-    partition_histogram = tl.zeros((BLOCK_NUM_EXPERTS,), dtype=tl.int32)
-    tile_offsets = tl.arange(0, BLOCK_NUM_ROUTES)
-
-    for route_start in tl.range(
-        partition_start, partition_end, BLOCK_NUM_ROUTES, loop_unroll_factor=1
-    ):
-        route_positions = route_start + tile_offsets
-        valid_routes = route_positions < partition_end
-        route_experts = tl.load(
-            flat_topk_indices + route_positions, mask=valid_routes, other=NUM_EXPERTS + tile_offsets
-        ).to(tl.int32)
-        ranks_in_tile = tl.inline_asm_elementwise(
-            asm="""
-            {
-                .reg .b32 matching_lanes;
-                .reg .b32 lower_lanes;
-                match.sync.any.b32 matching_lanes, $1, 0xffffffff;
-                mov.u32 lower_lanes, %lanemask_lt;
-                and.b32 matching_lanes, matching_lanes, lower_lanes;
-                popc.b32 $0, matching_lanes;
-            }
-            """,
-            constraints="=r,r",
-            args=[route_experts],
-            dtype=tl.int32,
-            is_pure=True,
-            pack=1,
-        )
-        safe_route_experts = tl.where(valid_routes, route_experts, 0)
-        first_warp_counts = tl.histogram(
-            route_experts, BLOCK_NUM_EXPERTS, mask=valid_routes & (tile_offsets < 32)
-        )
-        second_warp_counts = tl.histogram(
-            route_experts,
-            BLOCK_NUM_EXPERTS,
-            mask=valid_routes & (tile_offsets >= 32) & (tile_offsets < 64),
-        )
-        preceding_warp_counts = tl.gather(first_warp_counts, safe_route_experts, axis=0)
-        ranks_in_tile += tl.where(tile_offsets >= 32, preceding_warp_counts, 0)
-        ordinals_before_tile = tl.gather(partition_histogram, safe_route_experts, axis=0)
-        local_ordinals = ordinals_before_tile + ranks_in_tile
-        tl.store(
-            route_metadata + route_positions,
-            local_ordinals * BLOCK_NUM_EXPERTS + route_experts,
-            mask=valid_routes,
-        )
-        partition_histogram += first_warp_counts + second_warp_counts
-
-    tl.store(
-        partition_counts + partition * NUM_EXPERTS + expert_offsets,
-        partition_histogram,
-        mask=valid_experts,
-    )
-
-    _cooperative_grid_barrier(grid_sync, num_partitions)
-
-    # Transpose the prefix work: programs scan expert columns across route
-    # partitions and publish the exclusive offsets in place.
-    partition_offsets = tl.arange(0, BLOCK_SCAN_PARTITIONS)
-    valid_partitions = partition_offsets < num_partitions
-    for scan_expert_offset in tl.static_range(0, NUM_SCAN_EXPERTS):
-        scan_expert = partition + scan_expert_offset * num_partitions
-        valid_scan = valid_partitions & (scan_expert < NUM_EXPERTS)
-        counts_for_expert = tl.load(
-            partition_counts + partition_offsets * NUM_EXPERTS + scan_expert,
-            mask=valid_scan,
-            other=0,
-        )
-        tl.store(
-            partition_counts + partition_offsets * NUM_EXPERTS + scan_expert,
-            tl.cumsum(counts_for_expert, axis=0) - counts_for_expert,
-            mask=valid_scan,
-        )
-
-
 def _launch_route_ranking(
     workspace: ReplicaPlannerWorkspace,
     flat_topk_indices: torch.Tensor,
@@ -1917,163 +1603,15 @@ def _launch_route_ranking(
     num_experts: int,
     num_routes: int,
 ) -> None:
-    """Launch the one-kernel stable per-expert route ranking.
-
-    The intra-tile ranking reads ``match.sync`` lane masks and splits its
-    histogram into a first and a second warp, so the tile must be exactly two
-    warps wide: ``BLOCK_NUM_ROUTES`` and ``num_warps`` below are a correctness
-    constraint, not a tuning choice.
-    """
-    num_programs = _route_partition_count(num_routes)
-    _rank_routes_within_experts_kernel[(num_programs,)](
+    """Launch stable per-expert route ranking into the planner workspace."""
+    launch_replica_route_ranking(
         flat_topk_indices,
         workspace.sort_route_metadata,
         workspace.sort_partition_counts,
         workspace.sort_grid_sync,
-        NUM_ROUTES=num_routes,
-        NUM_EXPERTS=num_experts,
-        BLOCK_NUM_EXPERTS=triton.next_power_of_2(num_experts),
-        BLOCK_NUM_ROUTES=64,
-        BLOCK_SCAN_PARTITIONS=_next_power_of_two(num_programs),
-        NUM_SCAN_EXPERTS=triton.cdiv(num_experts, num_programs),
-        launch_cooperative_grid=True,
-        num_warps=2,
+        num_experts=num_experts,
+        num_routes=num_routes,
     )
-
-
-@triton.jit
-def _map_virtual_experts_kernel(
-    route_metadata,
-    partition_counts,
-    destination_boundaries,
-    expert_replica_slots,
-    virtual_experts,
-    NUM_ROUTES: tl.constexpr,
-    NUM_EXPERTS: tl.constexpr,
-    NUM_EXPERTS_PER_GPU: tl.constexpr,
-    EP_SIZE: tl.constexpr,
-    BLOCK_NUM_EXPERTS: tl.constexpr,
-    BLOCK_NUM_ROUTES: tl.constexpr,
-    BLOCK_EP_SIZE: tl.constexpr,
-    LOG2_BLOCK_EP_SIZE: tl.constexpr,
-):
-    """Map every local route to its rank-major native or replica expert id.
-
-    Routes for each expert form one deterministic global stream: source ranks
-    are concatenated in rank order, and routes within a source use flattened
-    token/top-k order. ``destination_boundaries`` holds that stream's partition
-    points already rebased into this rank's local ordinal space, so a route's
-    destination is simply the number of boundaries it has passed. Because the
-    boundaries are non-decreasing, the search is a branchless binary search
-    instead of a scan over every rank.
-
-    The grid matches the route-ranking partitioning so this kernel can finish
-    the ranking with a single indexed load, read its input sequentially, and
-    write ``virtual_experts`` in route order.
-    """
-    partition = tl.program_id(0)
-    num_partitions = tl.num_programs(0)
-    routes_per_partition = tl.cdiv(NUM_ROUTES, num_partitions)
-    partition_start = partition * routes_per_partition
-    partition_end = tl.minimum(partition_start + routes_per_partition, NUM_ROUTES)
-    expert_offsets = tl.arange(0, BLOCK_NUM_EXPERTS)
-    routes_before_partition = tl.load(
-        partition_counts + partition * NUM_EXPERTS + expert_offsets,
-        mask=expert_offsets < NUM_EXPERTS,
-        other=0,
-    )
-    tile_offsets = tl.arange(0, BLOCK_NUM_ROUTES)
-
-    for route_start in tl.range(
-        partition_start, partition_end, BLOCK_NUM_ROUTES, loop_unroll_factor=1
-    ):
-        route_positions = route_start + tile_offsets
-        valid_routes = route_positions < partition_end
-        packed_metadata = tl.load(route_metadata + route_positions, mask=valid_routes, other=0).to(
-            tl.int32
-        )
-        experts = packed_metadata % BLOCK_NUM_EXPERTS
-        ordinals_in_partition = packed_metadata // BLOCK_NUM_EXPERTS
-        safe_experts = tl.where(valid_routes, experts, 0)
-        # This route is the nth of its expert among all routes on this rank.
-        local_ordinal = (
-            tl.gather(routes_before_partition, safe_experts, axis=0) + ordinals_in_partition
-        )
-
-        # Count the destination segments this ordinal has already passed. The
-        # boundary row is padded up to BLOCK_EP_SIZE with the expert's local
-        # route total, which no ordinal reaches, so the count never exceeds the
-        # last real rank.
-        boundary_base = destination_boundaries + safe_experts * BLOCK_EP_SIZE
-        destination = tl.zeros((BLOCK_NUM_ROUTES,), dtype=tl.int32)
-        for step in tl.static_range(0, LOG2_BLOCK_EP_SIZE):
-            candidate = destination + (BLOCK_EP_SIZE >> (step + 1))
-            boundary = tl.load(boundary_base + candidate - 1)
-            destination = tl.where(local_ordinal >= boundary, candidate, destination)
-
-        # Native routes use the owner's local expert id. Remote routes look up
-        # the slot holding that expert's copied weights on the destination.
-        owner = experts // NUM_EXPERTS_PER_GPU
-        owned_local = experts % NUM_EXPERTS_PER_GPU
-        replica_slot = tl.load(
-            expert_replica_slots + safe_experts * EP_SIZE + destination,
-            mask=valid_routes & (destination != owner),
-            other=-1,
-        )
-        runtime_local = tl.where(
-            destination == owner, owned_local, NUM_EXPERTS_PER_GPU + replica_slot
-        )
-        virtual = destination.to(tl.int64) * (2 * NUM_EXPERTS_PER_GPU) + runtime_local
-        tl.store(virtual_experts + route_positions, virtual, mask=valid_routes)
-
-
-@triton.jit
-def _compact_routing_map_kernel(
-    routing_map,
-    token_indices,
-    tokens_per_expert,
-    num_tokens,
-    ROUTER_TOPK: tl.constexpr,
-    NUM_EXPERTS: tl.constexpr,
-    BLOCK_TOKENS: tl.constexpr,
-    BLOCK_NUM_EXPERTS: tl.constexpr,
-):
-    """Compact a dense routing map into per-token semantic expert ids.
-
-    The dense map already names every selected expert, so recovering the
-    compact route list is a per-row prefix sum rather than a selection
-    problem: the running count of set entries before an expert is the slot
-    that expert belongs in. The same pass accumulates the local per-expert
-    route histogram the planner gathers across the EP group.
-
-    The grid is persistent because the histogram is published with one atomic
-    update per program. Sizing the grid by tokens instead would make those
-    updates, not the map itself, the dominant cost.
-    """
-    program = tl.program_id(0)
-    experts = tl.arange(0, BLOCK_NUM_EXPERTS)
-    valid_experts = experts < NUM_EXPERTS
-    token_offsets = tl.arange(0, BLOCK_TOKENS)
-    histogram = tl.zeros((BLOCK_NUM_EXPERTS,), dtype=tl.int32)
-    tokens_per_program = tl.cdiv(num_tokens, tl.num_programs(0))
-    program_start = program * tokens_per_program
-    program_end = tl.minimum(program_start + tokens_per_program, num_tokens)
-
-    for token_start in tl.range(program_start, program_end, BLOCK_TOKENS, loop_unroll_factor=1):
-        tokens = token_start + token_offsets
-        valid = (tokens[:, None] < program_end) & valid_experts[None, :]
-        selected = tl.load(
-            routing_map + tokens[:, None] * NUM_EXPERTS + experts[None, :], mask=valid, other=0
-        ).to(tl.int32)
-        slots = tl.cumsum(selected, axis=1) - selected
-        tl.store(
-            token_indices + tokens[:, None] * ROUTER_TOPK + slots,
-            tl.broadcast_to(experts[None, :], (BLOCK_TOKENS, BLOCK_NUM_EXPERTS)),
-            mask=(selected != 0) & (slots < ROUTER_TOPK),
-        )
-        histogram += tl.sum(selected, axis=0)
-
-    tl.atomic_add(tokens_per_expert + experts, histogram, mask=valid_experts)
 
 
 def extract_semantic_routes(
@@ -2101,24 +1639,18 @@ def extract_semantic_routes(
     num_tokens, num_experts = (int(size) for size in routing_map.shape)
     device = routing_map.device
     tokens_per_expert = torch.zeros(num_experts, dtype=torch.int32, device=device)
-    if HAVE_TRITON and device.type == "cuda":
+    if device.type == "cuda":
         # Zeroed rather than empty: a routing map that selects fewer than
         # router_topk experts for some token would otherwise leave stale slots,
         # and the planner indexes tables with these ids.
         token_indices = torch.zeros((num_tokens, router_topk), dtype=torch.int32, device=device)
-        block_num_experts = _next_power_of_two(num_experts)
-        block_tokens = min(32, max(1, 16384 // block_num_experts))
-        num_programs = min(_MAX_PLANNER_PROGRAMS, triton.cdiv(num_tokens, block_tokens))
-        _compact_routing_map_kernel[(num_programs,)](
+        launch_compact_routing_map(
             routing_map,
             token_indices,
             tokens_per_expert,
-            num_tokens,
-            ROUTER_TOPK=router_topk,
-            NUM_EXPERTS=num_experts,
-            BLOCK_TOKENS=block_tokens,
-            BLOCK_NUM_EXPERTS=block_num_experts,
-            num_warps=8,
+            num_tokens=num_tokens,
+            router_topk=router_topk,
+            num_experts=num_experts,
         )
     else:
         # Reference formulation for CPU tests: the same prefix sum, expressed
@@ -2156,8 +1688,6 @@ def _validate_inputs(
         ``(num_tokens, router_topk, num_experts, ep_size)`` as Python integers
         for kernel launch configuration.
     """
-    if not HAVE_TRITON:
-        raise ImportError("The replica planner requires Triton.")
     if not dist.is_initialized():
         raise RuntimeError("The replica planner requires initialized torch.distributed.")
     if not topk_indices.is_cuda or not tokens_per_expert.is_cuda:
@@ -2215,26 +1745,19 @@ def _launch_replica_placement(
     num_local_experts: int,
 ) -> None:
     """Launch deterministic single-kernel replica placement."""
-    _plan_replica_placement_kernel[(ep_size,)](
+    launch_replica_placement(
         workspace.gathered_counts,
-        workspace.expert_totals,
         workspace.balance,
-        workspace.receiver_quotas,
         workspace.allocation,
         workspace.destination_boundaries,
         workspace.experts_to_copy,
         workspace.expert_replica_slots,
         workspace.placement_grid_sync,
-        RANK_ROUTE_CAPACITY=rank_route_capacity,
-        SOURCE_EP_RANK=source_rank,
-        EP_SIZE=ep_size,
-        NUM_EXPERTS=num_experts,
-        NUM_EXPERTS_PER_GPU=num_local_experts,
-        BLOCK_EP_SIZE=triton.next_power_of_2(ep_size),
-        BLOCK_NUM_EXPERTS_PER_GPU=triton.next_power_of_2(num_local_experts),
-        BLOCK_NUM_EXPERTS=triton.next_power_of_2(num_experts),
-        launch_cooperative_grid=True,
-        num_warps=1,
+        rank_route_capacity=rank_route_capacity,
+        source_rank=source_rank,
+        ep_size=ep_size,
+        num_experts=num_experts,
+        num_local_experts=num_local_experts,
     )
 
 
@@ -2247,22 +1770,16 @@ def _launch_replica_route_mapping(
     num_routes: int,
 ) -> None:
     """Map the ranked routes to native-or-replica ids."""
-    block_ep_size = triton.next_power_of_2(ep_size)
-    _map_virtual_experts_kernel[(_route_partition_count(num_routes),)](
+    launch_replica_route_mapping(
         workspace.sort_route_metadata,
         workspace.sort_partition_counts,
         workspace.destination_boundaries,
         workspace.expert_replica_slots,
         workspace.virtual_experts,
-        NUM_ROUTES=num_routes,
-        NUM_EXPERTS=num_experts,
-        NUM_EXPERTS_PER_GPU=num_local_experts,
-        EP_SIZE=ep_size,
-        BLOCK_NUM_EXPERTS=triton.next_power_of_2(num_experts),
-        BLOCK_NUM_ROUTES=256,
-        BLOCK_EP_SIZE=block_ep_size,
-        LOG2_BLOCK_EP_SIZE=block_ep_size.bit_length() - 1,
-        num_warps=8,
+        num_routes=num_routes,
+        num_experts=num_experts,
+        num_local_experts=num_local_experts,
+        ep_size=ep_size,
     )
